portage: Move the new util/futures.py to a futures subpkg
